Worked on event handlers, mostly.
Rolled PlayerPlaceBlockHandler and PlayerBreakBlockHandler into a
single PlayerPlaceOrBreakBlockHandler. It seems to make the most sense
to me as both are handled together in determining whether a player do
them.

Added Strings class as a central repository for message strings.

Added methods to Utils allowing it to determine the dimension a player
will spawn in and keeping track of whether a player has died and still
has to respawn. I may move that to a class for handling states? Idk.
Maybe if I have to do more of that.

Implemented PlayerDeathHandler.

Added messages that get sent to the player when they try to interact
with blocks or entities they shouldn’t be able to.

Added message if the player tries to enter a chunk they shouldn’t be
able to. Made sure to only apply this check when the player isn’t
currently respawning.

Added methods to the claim registry to check whether or not there are
claims with a given name or UUID.