Almost finished implementing ClaimRegistry
Only left out the file saving/loading functionality.

Added the ability to claim/unclaim chunks via the claim object.

Added methods to allow the the ClaimRegistry to add or remove chunks to
a claim in a way that doesn’t result in the claim then attempting the
same back on the ClaimRegistry.

Added a couple of exceptions.

Added ability to remove claim power from player even if that would
result in them having a negative claim power.